requirements.txt
crewai==0.35.8
crewai-tools==0.4.0
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
langchain-google-genai==1.0.1
langchain-community==0.0.38
python-multipart==0.0.9
